uploading dataset and tests from a paper I found online

+++ include/PCA.hpp
@@ -25,7 +25,7 @@ public :
 
         // get the sum of variances, this'll be useful later
         double sumVar = 0;
-        for (int i = 0; i < covariances.nRows(); i++) {
+        for (size_t i = 0; i < covariances.nRows(); i++) {
             sumVar += covariances(i, i);
         }
